WIP: MainWindow removal + additional mixins
* piekit.mainwindow -> MainWindow has been completely removed
* piekit.plugins.mixins -> Added new mixins
* piekit.utils.core -> Added new get_main_window method
* Changes in the PluginManager
* Fixes
